requirements.txt
pinecone-client==2.2.4
openai==1.12.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
pandas==2.1.4
numpy==1.24.3
tqdm==4.66.1 